Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@4682 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ GDiffView/src/nu/mine/mosher/Range.java
@@ -104,7 +104,7 @@ public class Range implements Comparable
     /**
      * @see java.lang.Comparable#compareTo(java.lang.Object)
      */
-    public long compareTo(Object o)
+    public int compareTo(Object o)
     {
         Range that = (Range)o;
         if (this.begin < that.begin)
